+++ java/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>org.simpledao</groupId>
     <artifactId>simpledao</artifactId>
-    <version>2.9.18</version>
+    <version>2.9.18-test</version>
     <name>SimpleDAO</name>
 
     <properties>
@@ -28,6 +28,20 @@
     </distributionManagement>
 
     <dependencies>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>6.2.9</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>26.0.2</version> <!-- Use the latest stable version -->
+            <scope>provided</scope>
+        </dependency>
 
         <dependency>
             <groupId>org.apache.commons</groupId>
