+++ java/src/org/simpledao/SimpleDAORepository.java
@@ -1,0 +1,58 @@
+package org.simpledao;
+
+import org.springframework.stereotype.Repository;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.ArrayList;
+
+@Repository
+public class SimpleDAORepository extends SimpleDAOBase
+{
+    private DataSource dataSource;
+
+    public SimpleDAORepository(DataSource dataSource) {
+        // Default constructor for Spring
+        this.dataSource = dataSource;
+    }
+
+    public <T> void simpleInsert( T bean ) throws SQLException
+    {
+        try (Connection con = new SimpleDBConnection().getDBConnection())
+        {
+            simpleInsert( con, bean, getBeanDescriptor(bean) );
+        }
+    }
+
+    public <T> void simpleUpdate( T bean ) throws SQLException
+    {
+        try (Connection con = new SimpleDBConnection().getDBConnection())
+        {
+            simpleUpdate( con, bean, getBeanDescriptor(bean) );
+        }
+    }
+
+    public <T> T simpleSelect(T criteria) throws SQLException
+    {
+        try (Connection con = new SimpleDBConnection().getDBConnection())
+        {
+            return simpleSelect(con, criteria, getBeanDescriptor(criteria));
+        }
+    }
+
+    public <T> ArrayList<T> simpleSelectList(T criteria ) throws SQLException
+    {
+        SimpleDBConnection dbc = new SimpleDBConnection();
+        Connection con = null;
+        try
+        {
+            con = dbc.getDBConnection();
+            return simpleSelectList( con, criteria);
+        }
+        finally
+        {
+            dbc.closeDBConnection(con);
+        }
+    }
+}
